git commit -m "Feat: 파티 엔티티에 isRecruiting 메서드 추가"

+++ matjalalzz/src/main/java/shop/matjalalzz/party/app/PartyService.java
@@ -4,7 +4,6 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
@@ -28,23 +27,21 @@ import shop.matjalalzz.party.dto.PartySearchCondition;
 import shop.matjalalzz.party.entity.Party;
 import shop.matjalalzz.party.entity.PartyUser;
 import shop.matjalalzz.party.entity.enums.GenderCondition;
-import shop.matjalalzz.party.entity.enums.PartyStatus;
 import shop.matjalalzz.party.mapper.PartyMapper;
 import shop.matjalalzz.shop.app.ShopService;
 import shop.matjalalzz.shop.entity.Shop;
-import shop.matjalalzz.user.dao.UserRepository;
+import shop.matjalalzz.user.app.UserService;
 import shop.matjalalzz.user.entity.User;
 
 @Service
 @RequiredArgsConstructor
-@Slf4j
 public class PartyService {
 
     private final PartyRepository partyRepository;
-    private final UserRepository userRepository;
     private final PartyUserRepository partyUserRepository;
     private final PartySchedulerService partySchedulerService;
     private final ShopService shopService;
+    private final UserService userService;
 
     @Transactional
     public void createParty(PartyCreateRequest request, long userId) {
@@ -55,7 +52,7 @@ public class PartyService {
 
         Party party = PartyMapper.toEntity(request, shop);
 
-        PartyUser host = PartyUser.createHost(party, getUserById(userId));
+        PartyUser host = PartyUser.createHost(party, userService.getUserById(userId));
         party.getPartyUsers().add(host);
 
         partyRepository.save(party);
@@ -112,25 +109,30 @@ public class PartyService {
     @Transactional
     public void joinParty(Long partyId, long userId) {
         Party party = findById(partyId);
-        User user = getUserById(userId);
+        User user = userService.getUserById(userId);
 
         validateJoinParty(party, user);
 
         HandlePartyUserJoin(party, user);
     }
 
-
     //TODO: 예약금 차감에 대한 로직 필요 (totalReservationFee 내려줘야함)
     @Transactional
     public void quitParty(Long partyId, long userId) {
         Party party = findById(partyId);
-        getUserById(userId); //검증용
+        userService.getUserById(userId); //검증용
         PartyUser partyUser = findPartyUser(userId, party);
 
         // 호스트인 경우 파티 탈퇴 불가능
         if (partyUser.isHost()) {
             throw new BusinessException(ErrorCode.HOST_CANNOT_QUIT_PARTY);
         }
+
+        // 모집 중인 파티만 탈퇴 가능
+        if (!party.isRecruiting()) {
+            throw new BusinessException(ErrorCode.CANNOT_QUIT_PARTY_STATUS);
+        }
+
         partyUser.delete();
         party.decreaseCurrentCount();
     }
@@ -139,28 +141,30 @@ public class PartyService {
     @Transactional
     public void deleteParty(Long partyId, long userId) {
         Party party = findById(partyId);
-        getUserById(userId); //검증용
+        userService.getUserById(userId); //검증용
         PartyUser partyUser = findPartyUser(userId, party);
 
-        if (partyUser.isHost()) { // 호스트인 경우만 파티 삭제 가능
-            if (!party.getStatus().equals(PartyStatus.RECRUITING)) { // 모집 중인 파티만 삭제 가능
-                throw new BusinessException(ErrorCode.CANNOT_DELETE_PARTY_STATUS);
-            }
-            party.deleteParty(); //파티 유저까지 cascade 삭제
-        } else {
+        // 호스트인 경우만 파티 삭제 가능
+        if (!partyUser.isHost()) {
             throw new BusinessException(ErrorCode.FORBIDDEN_ACCESS_DELETE_PARTY);
         }
+
+        // 모집 중인 파티만 삭제 가능
+        if (!party.isRecruiting()) {
+            throw new BusinessException(ErrorCode.CANNOT_DELETE_PARTY_STATUS);
+        }
+        party.deleteParty(); //파티 유저와 댓글까지 cascade 삭제
     }
 
     @Transactional
     public void completePartyRecruit(Long partyId, long userId) {
         Party party = findById(partyId);
 
-        if (!party.getStatus().equals(PartyStatus.RECRUITING)) {
+        if (!party.isRecruiting()) {
             throw new BusinessException(ErrorCode.ALREADY_COMPLETE_PARTY);
         }
 
-        getUserById(userId); //검증용
+        userService.getUserById(userId); //검증용
         PartyUser partyUser = findPartyUser(userId, party);
 
         // 호스트인 경우만 파티 상태 변경 가능
@@ -173,7 +177,7 @@ public class PartyService {
 
     private void validateJoinParty(Party party, User user) {
         // 1. 모집 상태 확인
-        if (!party.getStatus().equals(PartyStatus.RECRUITING)) {
+        if (!party.isRecruiting()) {
             throw new BusinessException(ErrorCode.NOT_RECRUITING_PARTY);
         }
 
@@ -228,6 +232,7 @@ public class PartyService {
             PartyUser partyUser = PartyUser.createUser(party, user);
             party.getPartyUsers().add(partyUser);
         }
+
         party.increaseCurrentCount();
     }
 
@@ -249,8 +254,4 @@ public class PartyService {
         return partyUserRepository.findAllByPartyId(partyId);
     }
 
-    private User getUserById(Long userId) {
-        return userRepository.findById(userId).orElseThrow(() ->
-            new BusinessException(ErrorCode.USER_NOT_FOUND));
-    }
 }
